[FUEGO] Several issues: - Change spec names in mechanism.h to comply with cpp standards: +/- are   p/n now. Several other issues might arise, it will be case by case - Adapt fuego so that non integer stoich coeffs are recognized. Seems to   be working ok on examples I generated. LiDryer ReactEval test case gives the same results after the changes than before the changes. - Adapt fuego to REV reactions (but needs to be tested)

+++ Support/Fuego/Pythia/pythia-0.4/packages/fuego/fuego/serialization/chemkin/unpickle/tokens/Reaction.py
@@ -157,7 +157,7 @@ class Reaction(Token):
 
         # Convert coefficient to an integer
         try:
-            c = int(coefficient)
+            c = float(coefficient)
         except TypeError: # attempt to int(None); set to 1
             c = 1
         except ValueError:
